GraphScope on 3/3 part 2: pick up the previous graphscope functions (#238)
* Pick up the previous graphscope functions
* Remove the usages of server_own
* Fixes the dataset iteration when one client has multiple servers
* Avoid print the log at the 0 step
* Add __init__.py and move versions to version.py
* Initializing all datasets at the start

+++ graphlearn/src/common/base/progress.h
@@ -55,7 +55,7 @@ struct Progress {
     if (count > stage * grading) {
       char buffer[100];
       snprintf(buffer, sizeof(buffer),
-               "Progress of %s: %ld", key.c_str(), stage * grading);
+               "Progress of %s: %lld", key.c_str(), (long long)(stage * grading));
       USER_LOG(buffer);
       ++stage;
     }
